contributors/ttwf/bensheldon/incoming/medialist-interfaces-001.html

<!DOCTYPE html>
<html>
 <head>
  <title>CSS Test: CSSOM Media Query List Serialization</title>
  <link rel="author" title="Ben Sheldon" href="mailto:ben@codeforamerica.org">
  <link rel="author" title="Chapman Shoop" href="mailto:chapman.shoop@gmail.com">
  <link rel="help" href="http://www.w3.org/TR/cssom/#the-medialist-interface">
  <meta name="flags" content="dom">
  <meta name="assert" content="MediaList objects have the properties specified in their interfaces">
  <script src="/resources/testharness.js" type="text/javascript"></script>
  <script src="/resources/testharnessreport.js" type="text/javascript"></script>
  
  <style id="styleElement" type="text/css" media="all" title="internal style sheet" disabled="disabled">
    * { margin: 0; padding: 0; }
  </style>

 </head>
 <body>
  <noscript>Test not run - javascript required.</noscript>
  <div id="log"></div>
  <script type="text/javascript">
  
    var styleElement = document.getElementById("styleElement");
    var styleSheet = styleElement.sheet;
    var mediaList = styleSheet.media;
  
    test(function() {
        assert_equals(mediaList.mediaText, "all");        
    }, "mediatest_medialist_serialize_element",
    { help: ["http://www.w3.org/TR/cssom/#the-medialist-interface", "http://www.w3.org/TR/cssom/#serializing-media-queries"],
      assert: ["MediaList.mediaText equals the 'media' value of the initial 'style' element."] });

    test(function() {      
      mediaList.appendMedium('screen');
      assert_equals(mediaList.mediaText, "all, screen");      
      
      // post-test cleanup
      mediaList.deleteMedium('screen');
        
    }, "mediatest_medialist_serialize_comma",
    { help: ["http://www.w3.org/TR/cssom/#the-medialist-interface", "http://www.w3.org/TR/cssom/#serializing-media-queries"],
      assert: ["To serialize a comma-separated list concatenate all items of the list in list order while separating them by \",\" (U+002C), followed by a space (U+0020)."] });
      
    test(function() {
      mediaList.deleteMedium('all');
      assert_equals(mediaList.mediaText, "");    
      
      // post-test cleanup
      mediaList.appendMedium('all');

    }, "mediatest_medialist_serialize_empty",
    { help: ["http://www.w3.org/TR/cssom/#the-medialist-interface", "http://www.w3.org/TR/cssom/#serializing-media-queries"],
      assert: ["If the media query list is empty return the empty string."] });
      
    test(function() {
      mediaList.appendMedium('screen');
      mediaList.appendMedium('print');
      assert_equals(mediaList.mediaText, "all, print, screen");    

    }, "mediatest_medialist_serialize_lexicographical",
    { help: ["http://www.w3.org/TR/cssom/#the-medialist-interface", "http://www.w3.org/TR/cssom/#serializing-media-queries"],
      assert: ["Each media query in the list of media queries should be sorted in lexicographical order."] });

  </script>
 </body>
</html>
